Proyecto_PNL/requirements.txt
Django>=4.2
djangorestframework>=3.15
channels>=4.0
sympy>=1.12
numpy>=1.26
python-dotenv>=1.0
groq>=0.10.0

# Opcionales (habilitan funciones adicionales del MVP)
# plotly>=5.24
# matplotlib>=3.9
# reportlab>=4.2
# scipy>=1.11
# channels-redis>=4.2
# osqp>=0.6
# quadprog>=0.1
# cvxopt>=1.3
